Fix kernel bugs from audit report (Batch 3)
This patch addresses several issues identified in the audit report:

1.  **Bug 59 (ioctl buffer check):** In `src/system/kernel/fs/fd.cpp`, updated `_user_ioctl` to use `is_user_address_range` for buffer validation instead of `IS_KERNEL_ADDRESS`. This ensures that user-provided buffers (pointer + length) are entirely within user address space, preventing potential kernel memory corruption or info leaks if a buffer spans across the boundary.

2.  **Bug 52 (RingBuffer race):** In `src/system/kernel/fs/fifo.cpp`, `RingBuffer::Read` now uses `atomic_get` to load `fWriteAvailable`. This prevents a potential race condition where the reader could observe an inconsistent state (new write head but old write available count) due to CPU or compiler reordering, leading to incorrect "empty buffer" detection.

3.  **Bug 57 (Signal to Kernel Thread):** In `src/system/kernel/fs/fifo.cpp`, `FIFOInode::Write` now checks the `isUser` flag before sending `SIGPIPE`. This prevents sending signals to kernel threads that might be writing to a pipe, avoiding potential unexpected behavior in kernel context.

4.  **Bug 56 (UserNodeListener RTTI):** In `src/system/kernel/fs/node_monitor.cpp` and `headers/private/kernel/Notifications.h`, replaced `dynamic_cast` with a virtual method `AsUserMessagingListener()` and sender verification. This removes the dependency on RTTI for identifying `UserNodeListener` objects, improving kernel safety and compatibility with RTTI-disabled builds.

Fixes part of Batch 3 from audit_report.md.

+++ src/system/kernel/util/ring_buffer.cpp
@@ -41,7 +41,7 @@ read_from_buffer(struct ring_buffer *buffer, uint8 *data, ssize_t length,
 {
 	size_t available = buffer->in;
 
-	if (length > (ssize_t)available)
+	if ((size_t)length > available)
 		length = available;
 
 	if (length == 0)
@@ -86,7 +86,7 @@ write_to_buffer(struct ring_buffer *buffer, const uint8 *data, ssize_t length,
 	bool user)
 {
 	size_t left = space_left_in_buffer(buffer);
-	if (length > (ssize_t)left)
+	if ((size_t)length > left)
 		length = left;
 
 	if (length == 0)
@@ -345,10 +345,10 @@ size_t
 ring_buffer_move(struct ring_buffer *to, ssize_t length,
 	struct ring_buffer *from)
 {
-	if (length > (ssize_t)from->in)
+	if ((size_t)length > (size_t)from->in)
 		length = from->in;
 
-	if (length > (ssize_t)(to->size - to->in))
+	if ((size_t)length > to->size - to->in)
 		length = to->size - to->in;
 
 	size_t bytesMoved = 0;
